separating executeScript
